Update example code in timeseries_dataset.py
I have gone through the example-2 in the TF documentation of keras.utils.timeseries_dataset_from_array. With existing code we can only generate one batch of dataset.In TF forum one of the user raised concern that users may confuse why we are using input_data = data[:-10] as it can generate only one batch and there will be loss of data. Though the intention in the example is to demo on generating batches of data using the API,Its better to use input_data = data[:] to avoid confusion and also it give perfect demo to generate the total possible no of batches without loss of data.

Also I am adding the sequence_length as variable in the code to avoid hard coding in the data indexing which will make it better understandable.

Thanks!