Refactor assert_no_new_pyobjects_executing_eagerly into a decorator factory.
Previously, `test_util.assert_no_new_pyobjects_executing_eagerly` could be used
either as a decorator or as a function which returned a decorator.

This change removes the (optional) `func` argument from the function and changes
the behavior such that it returns a decorator in all cases. This effectively transforms the
function into a pure decorator factory.

This change also updates uses of the decorator to all use the decorator factory form
(`@assert_no_new_pyobjects_executing_eagerly()`).

PiperOrigin-RevId: 583214989