Critical typo: squeeze the y_pred tensor even when it’s the same rank and shape as the y_pred tensor should be == instead of !=
Current behavior will squeeze the y_pred tensor even when it’s the same rank and shape as the y_pred tensor.

The issue is that the code is written wrong. Read the code comments and what it’s supposed to do (only squeeze when the ranks differ by exactly 1 and not squeeze for situations of equal rank). Yet the code’s if statements will squeeze whenever the ranks are equal, which is wrong from what I can tell.

the line: if (y_pred_rank - y_true_rank != 1) or y_pred_shape[-1] == 1: 
should be:
if (y_pred_rank - y_true_rank == 1) or y_pred_shape[-1] == 1: