Make `UnitNormalization` layer stateless.
There is no need to resolve negative axes in `build`, as `tf.linalg.l2_normalize` can handle them. Kept the build method to validate the axes in the context of the `input_shape`. Also added call to `super.build(...)` per best practice on Keras 2.

Note that in Keras 3, `UnitNormalization` is already stateless.

PiperOrigin-RevId: 699285760